Moved everything to the main project
