--- elasticsearch_service/go.mod
@@ -1,8 +1,0 @@
-module internship_project/elasticsearch_service
-
-go 1.15
-
-require (
-	github.com/segmentio/kafka-go v0.4.8
-	github.com/elastic/go-elasticsearch/v8 v8.0.0-20201104130540-2e1f801663c6
-	)
