--- elasticsearch_service/helpers/ElasticsearchClient.go
@@ -1,77 +1,0 @@
-package helpers
-
-import (
-	"context"
-	"encoding/json"
-	"github.com/elastic/go-elasticsearch/v8"
-	"github.com/elastic/go-elasticsearch/v8/esapi"
-	"log"
-	"strings"
-)
-
-type ElasticsearchClient struct {
-	client *elasticsearch.Client
-}
-
-func GetElasticsearchClient() ElasticsearchClient {
-	cfg := elasticsearch.Config{
-		Addresses: []string{
-			"http://localhost:9200",
-		},
-	}
-	es, err := elasticsearch.NewClient(cfg)
-	if err != nil {
-		log.Fatalf("Error getting client: %s", err)
-	}
-	return ElasticsearchClient{
-		client: es,
-	}
-}
-
-func (esclient *ElasticsearchClient) IndexDocument(id string, body string) {
-	req := esapi.IndexRequest{
-		Index:      "product",
-		DocumentID: id,
-		Body:       strings.NewReader(body),
-		Refresh:    "true",
-	}
-
-	res, err := req.Do(context.Background(), esclient.client)
-	if err != nil {
-		log.Fatalf("Error getting response: %s", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		log.Printf("[%s] Error indexing document ID=%s", res.Status(), id)
-	} else {
-		var r map[string]interface{}
-		if err := json.NewDecoder(res.Body).Decode(&r); err != nil {
-			log.Printf("Error parsing the response body: %s", err)
-		} else {
-			log.Printf("[%s] %s; version=%d", res.Status(), r["result"], int(r["_version"].(float64)))
-		}
-	}
-}
-
-func (esclient *ElasticsearchClient) DeleteDocument(id string) {
-	req := esapi.DeleteRequest{
-		Index: "product",
-		DocumentID: id,
-		Refresh: "true",
-	}
-	res, err := req.Do(context.Background(), esclient.client)
-	if err != nil {
-		log.Fatalf("Error getting response: %s", err)
-	}
-	defer res.Body.Close()
-	if res.IsError() {
-		log.Printf("[%s] Error deleting document ID=%s", res.Status(), id)
-	} else {
-		var r map[string]interface{}
-		if err := json.NewDecoder(res.Body).Decode(&r); err != nil {
-			log.Printf("Error parsing the response body: %s", err)
-		} else {
-			log.Printf("[%s] %s; version=%d", res.Status(), r["result"], int(r["_version"].(float64)))
-		}
-	}
-}
