--- elasticsearch_service/main.go
@@ -1,18 +1,0 @@
-package main
-
-import (
-	"internship_project/elasticsearch_service/helpers"
-	"sync"
-)
-
-var (
-	r  map[string]interface{}
-	wg sync.WaitGroup
-)
-
-func main() {
-	EsClient := helpers.GetElasticsearchClient()
-
-	kafkaConsumer := helpers.NewConsumer("ava-internship", EsClient)
-	kafkaConsumer.Consume()
-}
